Use sbatch on publish reconcile to improve pubperf [#12]
High I/O jobs will run slower on the head node because of the age of the
hardware. Compute nodes are harder, better, faster and stronger and will
perform these tasks better. This change sends the de novo FASTA/TSV
matching step to SLURM and blocks the publish script for the result.

I tested this last night and it halves the execution time for reconcile.
This is part of an upcoming suite of changes to improve the performance
of the post-Elan publishing process to cope with the scale we have now
reached at COG.