Remove cat consensus to improve pubperf [#12]
Elan has been generating an "unmatched" and "matched" FASTA file for the
past few months. The former is generated from `cat`'ing the symlinks of
the day into a new file; the latter is generated by the
reconcile_consensus script. Other than ordering and the tidying of some
FASTA headers, these are the same file. Same sequences.

The unmatched FASTA has always had an issue with the headers (ie. some
headers from early COG are dirty and I never went back to clean them).
Seeing as users are encouraged to use the matched FASTA and the headers
in the unmatched FASTA are unreliable, it seems unnecessary to maintain
it going forward. Removing the unmatched `cat` consensus means we save
approximately 20 minutes of file system wrangling.