DEPLOY THE CAFFEINE CAT [#D154,#O010]
Significantly improve performance of the post-Elan publish step

Unfortunately for both the UK and our cephfs deployment, our data set
continues to grow with rapid pace. The latest victim of the complexities
of scale is the 'post-Elan' publish step (cog-publish.sh).

This step conceptually `cat`'s all of the QC passing consensus sequences
together to "publish" the entire data set for the day. This rather
simple step has slowly but surely scaled in a linear fashion without too
much concern. However, a related cephfs scale issue (the "big fucking
dir problem") has recently destroyed the performance and predictability
of this publish step. Recent post-publish jobs have gone on longer than
eight hours, for sticking a bunch of files together.

On further investigation, the performance issues were attributed to tail
latency amplification of calls for files via the cephfs metadata server.
Essentially the cache is evicting the published directories because they
are too large and something fresher has bumped them out of memory. Queries
hitting the file metadata cache after the big dir has been evicted trigger
an expensive rescan of the entire big dir. While only a small and random
proportion of queries will trigger this eventuality, we make a substantial
number of queries while reconciling the daily consensus.

Long term these big published artifact directories will need to be sharded
but in the meantime this change achieves two goals:

    * Makes a future change to shard the published directory structure
      easier by not accessing previously stored objects directly
    * Improves performance of the existing system by two orders of
      magnitude, allowing us to tolerate the malfunctioning design long
      enough to fix it properly

This change improves the performance and predictability of the post-Elan
publish step by significantly reducing the number of file system queries.
This is done by suppressing a hit to the file system when a sequence falls
into one of the following scenarios:

    * The sequence does not exist
    * The sequence failed basic QC
    * The sequence was previously suppressed
    * The sequence was previously published

This is achieved by using the most recently published consensus as a
cache, addressed by the name of the Published Artifact Group (PAG). In
the case where the current PAG from an iterator over the M1 metadata table
is not in the current consensus cache it must either be new, or;
it does not exist, has been suppressed in the past, or is bad metadata with
no sequencing file (recall the M1 metadata table is not very clever).

We determine if the PAG is new by screening it against the recently
added "latest pass" file list. In all other cases we drop the sequence
without checking for it against the file system. To maintain existing
behaviour we also screen PAGs against the "latest kill" file list to ensure
that suppressed PAGs are dropped from the consensus.

Testing this change on the 20211118 data set yields a reconciled
consensus with the exact same MD5sum as the production consensus,
in just six minutes as opposed to an absolutely shocking eight hours.

This emergency change has been approved and will be deployed to execute
as part of tomorrow's Elan, 20211119.